+++ src/renderer/src/data/versionQuotes.ts
@@ -10,4 +10,5 @@
 export const versionQuotes: Record<string, string> = {
   '0.0.0': 'In the beginning, there was nothing',
   '0.1.0': 'You guys are getting interviews?',
+  '0.2.0': "We'll keep your resume on file",
 }